inventory_ai/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
alembic
pydantic
python-dotenv
pandas
prophet